Fixing validation of too many arguments in gprecoverseg
**Issue:**
When gprecoverseg is run with too many arguments then it errors out with an exception stack trace.

**RCA:**
Ideally when too many arguments are provided then gprecoverseg should raise an exception and error out gracefully.
There is stack trace seen because while handling this exception another exception occured where a non-existing attribute is accessed by a NoneType object.

**Fix:**
Instanting the object to a proper class so that it can access its attribute.